repo: Remove dynamic files to keep patching easier
The lockfile will get recreated as part of the build. However, as this
patchset will be ported forward to future versions of Rust, we want to
prevent subtle changes in the lockfile from allowing the patchset to
apply.

Therefore, remove the lockfile as the first commit.

We also remove the github CI file which changes a lot every release.

Signed-off-by: Sean Cross <sean@xobs.io>