libstd: xous: allocate stack + guard pages together
Perform one allocation for the thread stack and the guard pages, then
use the new `UpdateMemoryFlags` call to remove read and execute
permissions for the remaining pages. This prevents a race condition
where another thread can allocate memory between the stack and the guard
page.

As part of this patch, the libxous version was bumped to 0.9.7, which
made `MemoryRange::new()` an `unsafe` operation, which required a
modification to the `thread_local_key` implementation.

Signed-off-by: Sean Cross <sean@xobs.io>