Re-implement Shutdown, and Close on Drop for TcpStream
Shutdown does not Close a connection (counter-intuitively).

In fact, the semantics for Shutdown are to stop any tx or rx that
is currently waiting (so disregard the expiration timers) and
have them return Ok(0). It's a weird call.

The operation that was previously in Shutdown is migrated to
the Drop method, so that when the object goes out of scope the
file handles are de-allocated so they can be re-used.

Note that as part of this fix, we have to be able to return
a zero-length rx or tx, which is not possible with the encoding
used on the `valid` field (`NonZeroUsize`). So, we use the value
`u32::MAX` to represent zero. The alternative of striping the value
into the return memory is a big change, and also, the value of length
should be far, far away from u32::MAX (64k max TCP packet size but
also 1500 byte MTU means we won't collide with that encoding).