fix locking on dlmalloc
dlmalloc is decidely not thread-safe. Therefore it is mandatory to
have a big fat lock on all of its calls.

In the previous implementation, which was copied from `wasm`,
there is an option for a path that has atomics, and one for a path
that doesn't. The non-atomic path replaces the lock with a no-op.

It was assumed that whatever magic sets features would configure
atomics, and we'd be using the atomics path. Testing by deliberately
injecting failures into the two code paths empirically demonstrates
that we're using the non-atomics path, and the lock isn't happening.

A search through the code seems to reveal this is the only spot where
the "atomics" dependency is checked not in conjunction with a "wasm"
host check. Thus this is the only place (I think) inside the library
where this failure to configure atomics correctly actually has an
impact (in the rest of the code paths, the lack of atomics in the feature
config is ignored, and we use the version of the code that relies on
atomics).

So, it checks out that this incorrectly configured feature isn't
causing bigger problems.

However, there is really no reason for us to have both optional
code paths for Xous -- we always have atomics, and we always want
locking because we have threads. So, this commit removes even the
option of having the locks bypassed, and replaces them with a spin
lock that includes a call to "yield_slice()" when we're spinning.
There is also a call to yield_slice() on drop so as to ensure
that we don't starve out a waiting process by re-acquiring the
lock within our same scheduler quantum.

force noisy error if acquire_global_lock() is accidentally used

In theory this path should never be activated. Previously, it
would silently fail. Now, there is an `unimplemented!()` call
which should force a panic if we accidentally rely upon a lock
that is not implemented.