std: xous: add senres for more complex ffi
FFI between services is done with raw arrays of bytes. Senres was
developed as an easier way of dealing with these arrays.

Add an implementation of Senres that is compatible with Xous services
such as PDDB and the network server. It will be used to add filesystem
and networking support.

Signed-off-by: Sean Cross <sean@xobs.io>