bootstrap: builder: don't add origin paths on xous
Don't add the origin rpath when calling the linker. This is unnecessary
on Xous since there is no dynamic linker, and Xous can use an ordinary
bare-metal linker with no rpath options.

As part of this patch, the logic was inverted so that the "else" clause
contains everything except "windows" and "xous".

Signed-off-by: Sean Cross <sean@xobs.io>